Etape 9

+++ carburant.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include "robot.h"
 #include "traceur.h"
 #include "carburant.h"
@@ -6,7 +5,7 @@
 RobotCarburant::RobotCarburant(char aspect, char trace, int carbu) : RobotTraceur{aspect, trace}, m_carburant{carbu}
 {}
 
-void RobotCarburant::avancer(int distance)
+void RobotCarburant::avancer(double distance)
 {
   if(m_position < m_carburant)
     m_position += distance;
