+++ rapide.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include "robot.h"
 #include "rapide.h"
 
@@ -7,7 +6,7 @@ RobotRapide::RobotRapide(char aspect) : Robot{aspect}
 {}
 
 
-void RobotRapide::avancer(int distance)
+void RobotRapide::avancer(double distance)
 {
   m_position += distance *2;
 }
