+++ robot.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include "robot.h"
 
 
@@ -8,7 +7,7 @@ Robot::Robot(char aspect) : m_aspect{aspect}, m_position{0}
 Robot::~Robot(){}
 
 
-void Robot::avancer(int distance)
+void Robot::avancer(double distance)
 {
   m_position += distance;
 }
@@ -17,7 +16,7 @@ void Robot::montrer()
 {
   //for(int i = 0; i<m_position; i++)
   //{ *i
-      std::cout << std::string(m_position, ' ') << m_aspect << std::endl;
+  std::cout << std::string(m_position, ' ') << m_aspect << std::endl;
   //}
 
 }
