+++ teleporteur.cpp
@@ -1,4 +1,3 @@
-#include <iostream>
 #include "robot.h"
 #include "teleporteur.h"
 
@@ -7,7 +6,7 @@ RobotTeleporteur::RobotTeleporteur(char aspect) : Robot{aspect}
 {}
 
 
-void RobotTeleporteur::teleporter(int position)
+void RobotTeleporteur::teleporter(double position)
 {
   m_position = position;
 }
